requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.4
botocore==1.31.21
brotlipy==0.7.0
colorama==0.4.4
dataclasses-json==0.5.14
docutils==0.16
frozenlist==1.4.0
jmespath==1.0.1
langchain==0.0.294
langsmith==0.0.38
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
opensearch-py==2.3.1
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.5.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.6.2
SQLAlchemy==2.0.21
tenacity==8.2.3
traitlets==5.10.0
typing-inspect==0.9.0
wcwidth==0.2.6
yarl==1.9.2